requirements.txt
# Core LangChain
langchain==0.2.14
langchain-community==0.2.12

# Provider-specific integrations
langchain-openai==0.1.8
langchain-groq==0.1.4
langchain-ollama==0.1.3
langchain-huggingface==0.0.3

# Official SDKs
openai>=1.3.5,<2.0.0
groq>=0.4.2,<1.0.0
ollama>=0.3.0,<1.0.0

# Vector DBs and tokenizers
faiss-cpu>=1.8.0.post1,<2.0.0
chromadb>=0.5.3,<1.0.0
tiktoken>=0.7.0,<1.0.0

# HuggingFace models
torch>=2.2.0
transformers>=4.40.0

# Unstructured document loaders
unstructured[local-inference]>=0.15.0,<0.16.0
requests>=2.32.0,<3.0.0
beautifulsoup4>=4.12.0,<5.0.0
lxml>=5.2.2,<6.0.0
pdfminer.six>=20221105
pillow>=10.0.0
python-docx>=1.1.0
openpyxl>=3.1.2

# Utilities
protobuf>=4.25.0,<5.0.0
python-dotenv>=1.0.0,<2.0.0

streamlit==1.36.0
langchain-chroma==0.1.4
